+++ system/lib/actor_system.dart
@@ -1,4 +1,4 @@
-export 'src/system/actor.dart' show Actor, ActorFactory, ExternalActorCreate, ExternalActorLookup;
+export 'src/system/actor.dart' show Actor, ActorFactory, CreateActor, LookupActor;
 export 'src/system/base.dart' show actorScheme, localSystem, actorPath, localActorPath;
 export 'src/system/context.dart'
     show BaseContext, ActorContext, ActorSystem, MissingHostHandling, PathMatcher, patternMatcher;